// src/main/models/entity/master/MasterRole.ts
import { Sequelize } from 'sequelize';
import { Column, DataType, HasMany, Model, Table } from 'sequelize-typescript';
import UserRoles from '../credentials/UserRole';

export interface RolesAttributes {
  id: number;
  name: string;
  isHide?: boolean;
  createdBy: string;
  createdAt: Date;
  updatedAt?: Date;
  deletedAt?: Date;
}

type RolesAutoGeneratedAttributes =
  | 'id'
  | 'createdAt'
  | 'updatedAt'
  | 'deletedAt';
export type RolesCreationAttributes = Omit<
  RolesAttributes,
  RolesAutoGeneratedAttributes
>;

@Table({
  freezeTableName: true,
  tableName: 'roles',
  underscored: true,
  paranoid: true,
  schema: 'master',
})
class Roles
  extends Model<RolesAttributes | RolesCreationAttributes>
  implements RolesAttributes {
  @Column({
    allowNull: false,
    type: DataType.BIGINT,
    primaryKey: true,
    autoIncrement: true,
  })
  id!: number;

  @Column({
    allowNull: false,
    type: DataType.STRING,
    unique: true,
  })
  name!: string;

  @Column({
    allowNull: true,
    type: DataType.BOOLEAN,
    defaultValue: false
  })
  isHide!: boolean;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  createdBy!: string;

  @Column({
    allowNull: false,
    type: DataType.DATE,
    defaultValue: Sequelize.literal('now()'),
  })
  createdAt!: Date;

  @Column({
    type: DataType.DATE,
  })
  updatedAt!: Date;

  @Column({
    type: DataType.DATE,
  })
  deletedAt!: Date;

  @HasMany(() => UserRoles, {
    foreignKey: 'role_id',
    sourceKey: 'id',
  })
  user_roles!: UserRoles[];
}

export default Roles;