// src/main/models/entity/credentials/UserRole.ts
import { Sequelize } from 'sequelize';
import {
  BelongsTo,
  Column,
  DataType,
  ForeignKey,
  HasOne,
  Model,
  Table,
} from 'sequelize-typescript';
import Users from './Users';
import Roles from '../master/MasterRole';

export interface UserRolesAttributes {
  id: number;
  userId: string;
  roleId: number;
  createdAt: Date;
  updatedAt?: Date;
  deletedAt?: Date;
}

type UserRolesAutoGeneratedAttributes =
  | 'id'
  | 'createdAt'
  | 'updatedAt'
  | 'deletedAt';
export type UserRolesCreationAttributes = Omit<
  UserRolesAttributes,
  UserRolesAutoGeneratedAttributes
>;

@Table({
  freezeTableName: true,
  tableName: 'user_roles',
  paranoid: true,
  schema: 'credentials',
  underscored: true,
})
class UserRoles
  extends Model<UserRolesAttributes | UserRolesCreationAttributes>
  implements UserRolesAttributes {
  @Column({
    primaryKey: true,
    type: DataType.NUMBER,
    autoIncrement: true,
  })
  id!: number;

  @ForeignKey(() => Users)
  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  userId!: string;

  @ForeignKey(() => Roles)
  @Column({
    allowNull: false,
    type: DataType.NUMBER,
  })
  roleId!: number;

  @Column({
    allowNull: false,
    type: DataType.DATE,
    defaultValue: Sequelize.literal('now()'),
  })
  createdAt!: Date;

  @Column({
    type: DataType.DATE,
  })
  updatedAt!: Date;

  @Column({
    type: DataType.DATE,
  })
  deletedAt!: Date;


  @BelongsTo(() => Roles, {
    onDelete: 'cascade',
    onUpdate: 'cascade',
  })
  role!: Roles;

  @BelongsTo(() => Users, {
    onDelete: 'cascade',
    onUpdate: 'cascade',
  })
  user!: Users;
}

export default UserRoles;