// src/main/models/entity/master/Employee.ts
import { Sequelize } from 'sequelize';
import { BelongsTo, BelongsToMany, Column, DataType, ForeignKey, HasMany, Model, Table } from 'sequelize-typescript';
import sharp from 'sharp';
import Users from '../credentials/Users';

export interface EmployeeAttributes {
  id: string;
  foto: string;
  nik: string;
  fullname: string;
  noTelp?: string;
  noHp: string;
  email: string;
  address: string;
  coachId?: string;
  logHistory: Array<Object> | null;
  userId?: string;
  flag?: number;
  joinDate: Date;
  bornDate: Date;
  createdBy: string;
  createdAt: Date;
  updatedBy?: string;
  updatedAt?: Date;
  deletedAt?: Date;
  companyId?: string;
  companyInfo?: any;
}

type EmployeeAutoGeneratedAttributes =
  | 'id'
  | 'createdAt'
  | 'updatedAt'
  | 'deletedAt';
export type EmployeeCreationAttributes = Omit<
  EmployeeAttributes,
  EmployeeAutoGeneratedAttributes
>;

@Table({
  freezeTableName: true,
  tableName: 'employee',
  underscored: true,
  paranoid: true,
  schema: 'master',
})
class Employee
  extends Model<EmployeeAttributes | EmployeeCreationAttributes>
  implements EmployeeAttributes {

  @Column({
    allowNull: false,
    type: DataType.STRING,
    defaultValue: Sequelize.literal('uuid_generate_v7()'),
    primaryKey: true,
  })
  id!: string;

  @Column({
    allowNull: true,
    type: DataType.TEXT,
  })
  foto!: string;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  nik!: string;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  fullname!: string;

  @Column({
    allowNull: true,
    type: DataType.STRING(30),
  })
  noTelp!: string;

  @Column({
    allowNull: false,
    type: DataType.STRING(30),
  })
  noHp!: string;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  email!: string;

  @Column({
    allowNull: false,
    type: DataType.TEXT,
  })
  address!: string;

  @Column({
    allowNull: false,
    type: DataType.DATEONLY,
  })
  joinDate!: Date;

  @Column({
    allowNull: false,
    type: DataType.DATEONLY,
  })
  bornDate!: Date;

  @ForeignKey(() => Employee)
  @Column({
    allowNull: true,
    type: DataType.STRING,
  })
  coachId!: string;


  @ForeignKey(() => Users)
  @Column({
    allowNull: true,
    type: DataType.STRING,
  })
  userId!: string;

  @Column({
    allowNull: true,
    type: DataType.JSONB,
  })
  logHistory!: Array<Object> | null

  @Column({
    type: DataType.INTEGER,
    defaultValue: 1,
  })
  flag!: number;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  createdBy!: string;

  @Column({
    allowNull: false,
    type: DataType.DATE,
    defaultValue: Sequelize.literal('now()'),
  })
  createdAt!: Date;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  updatedBy!: string;

  @Column({
    type: DataType.DATE,
  })
  updatedAt!: Date;

  @Column({
    type: DataType.DATE,
  })
  deletedAt!: Date;

  @Column({
    allowNull: true,
    type: DataType.STRING,
  })
  companyId?: string;

  @Column({
    allowNull: true,
    type: DataType.JSONB,
  })
  companyInfo?: any;


  @BelongsTo(() => Users, {
    targetKey: "id",
    onDelete: 'cascade',
    onUpdate: 'cascade',
  })
  user!: Users;

}

export default Employee;