// src/main/models/entity/master/CompanyUser.ts
import { Sequelize } from 'sequelize';
import { BelongsTo, Column, DataType, ForeignKey, Model, Table } from 'sequelize-typescript';
import Company from './Company';
import Users from '../credentials/Users';

export interface CompanyUserAttributes {
    id: string;
    companyId: string;
    userId: string;
    createdBy: string;
    createdAt: Date;
    updatedBy?: string;
    updatedAt?: Date;
    deletedAt?: Date;
}

type CompanyUserAutoGeneratedAttributes =
    | 'id'
    | 'createdAt'
    | 'updatedAt'
    | 'deletedAt';

export type CompanyUserCreationAttributes = Omit<
    CompanyUserAttributes,
    CompanyUserAutoGeneratedAttributes
>;

@Table({
    freezeTableName: true,
    tableName: 'company_user',
    underscored: true,
    paranoid: true,
    schema: 'master',
})
class CompanyUser
    extends Model<CompanyUserAttributes | CompanyUserCreationAttributes>
    implements CompanyUserAttributes {
    @Column({
        allowNull: false,
        type: DataType.STRING,
        defaultValue: Sequelize.literal('uuid_generate_v7()'),
        primaryKey: true,
    })
    id!: string;

    @ForeignKey(() => Company)
    @Column({
        allowNull: false,
        type: DataType.STRING,
    })
    companyId!: string;


    @ForeignKey(() => Users)
    @Column({
        allowNull: false,
        type: DataType.STRING,
    })
    userId!: string;

    @Column({
        allowNull: false,
        type: DataType.STRING,
    })
    createdBy!: string;

    @Column({
        allowNull: false,
        type: DataType.DATE,
        defaultValue: Sequelize.literal('now()'),
    })
    createdAt!: Date;

    @Column({
        allowNull: true,
        type: DataType.STRING,
    })
    updatedBy!: string;

    @Column({
        type: DataType.DATE,
    })
    updatedAt!: Date;

    @Column({
        type: DataType.DATE,
    })
    deletedAt!: Date;

    @BelongsTo(() => Company)
    company!: Company;

    @BelongsTo(() => Users)
    user!: Users;

}

export default CompanyUser;