// src/main/models/entity/master/Company.ts
import { Sequelize } from 'sequelize';
import { BelongsToMany, Column, DataType, HasMany, Model, Table } from 'sequelize-typescript';
import { CompanyStatusList } from '../../../const';
import Users from '../credentials/Users';
import CompanyUser from './CompanyUser';



export interface CompanyAttributes {
    id: string;
    label: string;
    address: string;
    email: string;
    noTelp: string;
    status?: typeof CompanyStatusList[number];
    maxUser?: number;
    cost?: number;
    costTotal?: number;
    startDate?: Date;
    endDate?: Date;
    remainderDate?: Date;
    createdBy: string;
    createdAt: Date;
    updatedBy?: string;
    updatedAt?: Date;
    deletedAt?: Date;
}

type CompanyAutoGeneratedAttributes =
    | 'id'
    | 'createdAt'
    | 'updatedAt'
    | 'deletedAt';

export type CompanyCreationAttributes = Omit<
    CompanyAttributes,
    CompanyAutoGeneratedAttributes
>;

@Table({
    freezeTableName: true,
    tableName: 'company',
    underscored: true,
    paranoid: true,
    schema: 'master',
})
class Company
    extends Model<CompanyAttributes | CompanyCreationAttributes>
    implements CompanyAttributes {
    @Column({
        allowNull: false,
        type: DataType.STRING,
        defaultValue: Sequelize.literal('uuid_generate_v7()'),
        primaryKey: true,
    })
    id!: string;

    @Column({
        allowNull: false,
        type: DataType.STRING,
    })
    code!: string;

    @Column({
        allowNull: false,
        type: DataType.STRING,
    })
    label!: string;

    @Column({
        allowNull: false,
        type: DataType.TEXT,
    })
    address!: string;

    @Column({
        allowNull: false,
        type: DataType.STRING,
    })
    email!: string;

    @Column({
        allowNull: false,
        type: DataType.STRING(30),
    })
    noTelp!: string;

    @Column({
        allowNull: true,
        type: DataType.ENUM(...CompanyStatusList),
        defaultValue: 'PENDING',
    })
    status!: typeof CompanyStatusList[number];

    @Column({
        allowNull: true,
        type: DataType.INTEGER,
    })
    maxUser!: number;

    @Column({
        allowNull: true,
        type: DataType.DECIMAL(32, 16),
    })
    cost!: number;

    @Column({
        allowNull: true,
        type: DataType.DECIMAL(32, 16),
    })
    costTotal!: number;

    @Column({
        allowNull: true,
        type: DataType.DATEONLY,
    })
    startDate!: Date;

    @Column({
        allowNull: true,
        type: DataType.DATEONLY,
    })
    endDate!: Date;

    @Column({
        allowNull: true,
        type: DataType.DATEONLY,
    })
    remainderDate!: Date;

    @Column({
        allowNull: false,
        type: DataType.STRING,
    })
    createdBy!: string;

    @Column({
        allowNull: false,
        type: DataType.DATE,
        defaultValue: Sequelize.literal('now()'),
    })
    createdAt!: Date;

    @Column({
        allowNull: true,
        type: DataType.STRING,
    })
    updatedBy!: string;

    @Column({
        type: DataType.DATE,
    })
    updatedAt!: Date;

    @Column({
        type: DataType.DATE,
    })
    deletedAt!: Date;

    @HasMany(() => CompanyUser, {
        foreignKey: 'companyId',
        sourceKey: 'id',
        onDelete: 'cascade',
        onUpdate: 'cascade',
    })
    companyUser!: CompanyUser[];
}

export default Company;